handling aliases

+++ WebApplication/Processing/Publisher.cs
@@ -63,10 +63,17 @@ namespace WebApplication.Processing
                 throw new Exception($"App Bundle with package is not found ({packagePathname}).");
 
             try {
+                // checking existance of the AppBundle
+                await _client.GetAppBundleVersionsAsync(config.Bundle.Id);
+                Alias oldVersion = null;
+                try {
+                    // getting version of AppBundle for the particular Alias (label)
+                    oldVersion = await _client.GetAppBundleAliasAsync(config.Bundle.Id, config.Label);
+                } catch (System.Net.Http.HttpRequestException e) when (e.Message.Contains("404")) {};
                 Trace($"Creating new app bundle '{config.Bundle.Id}' version.");
-                Alias oldVersion = await _client.GetAppBundleAliasAsync(config.Bundle.Id, config.Label);
                 await _client.UpdateAppBundleAsync(config.Bundle, config.Label, packagePathname);
-                await _client.DeleteAppBundleVersionAsync(config.Bundle.Id, oldVersion.Version);
+                if (oldVersion != null)
+                    await _client.DeleteAppBundleVersionAsync(config.Bundle.Id, oldVersion.Version);
             }
             catch (System.Net.Http.HttpRequestException e) when (e.Message.Contains("404"))
             {
@@ -96,10 +103,17 @@ namespace WebApplication.Processing
             };
 
             try {
+                // checking existance of Activity
+                await _client.GetActivityVersionsAsync(config.ActivityId);
+                Alias oldVersion = null;
+                try {
+                    // getting version of Activity for the particular Alias (label)
+                    oldVersion = await _client.GetActivityAliasAsync(config.ActivityId, config.ActivityLabel);
+                } catch (System.Net.Http.HttpRequestException e) when (e.Message.Contains("404")) {};
                 Trace($"Creating new activity '{config.ActivityId}' version.");
-                Alias oldVersion = await _client.GetActivityAliasAsync(config.ActivityId, config.ActivityLabel);
                 await _client.UpdateActivityAsync(activity, config.ActivityLabel);
-                await _client.DeleteActivityVersionAsync(config.ActivityId, oldVersion.Version);
+                if (oldVersion != null)
+                    await _client.DeleteActivityVersionAsync(config.ActivityId, oldVersion.Version);
             }
             catch (System.Net.Http.HttpRequestException e) when (e.Message.Contains("404"))
             {
